Split `HasLink::class` and `HasLinkAttribute::class`.
