+++ src/Attribute/Component/HasLink.php
@@ -4,17 +4,12 @@ declare(strict_types=1);
 
 namespace PHPForge\Html\Attribute\Component;
 
-use PHPForge\Html\Helper\CssClass;
-
-use function array_merge;
-
 /**
- * Is used by widgets that implement the link methods.
+ * Is used by widgets that implement the link method.
  */
 trait HasLink
 {
     protected string $link = '';
-    protected array $linkAttributes = [];
 
     /**
      * @return string The link of the menu item.
@@ -22,14 +17,6 @@ trait HasLink
     public function getLink(): string
     {
         return $this->link;
-    }
-
-    /**
-     * @return array The `HTML` attributes of the href of the menu item.
-     */
-    public function getLinkAttributes(): array
-    {
-        return $this->linkAttributes;
     }
 
     /**
@@ -43,36 +30,6 @@ trait HasLink
     {
         $new = clone $this;
         $new->link = $value;
-
-        return $new;
-    }
-
-    /**
-     * Set the `HTML` attributes for the link tag.
-     *
-     * @param array $values Attribute values indexed by attribute names.
-     *
-     * @return static A new instance of the current class with the specified link attributes.
-     */
-    public function linkAttributes(array $values): static
-    {
-        $new = clone $this;
-        $new->linkAttributes = array_merge($this->linkAttributes, $values);
-
-        return $new;
-    }
-
-    /**
-     * Set the `CSS` class for the link tag.
-     *
-     * @param string $value The `CSS` class for the link tag.
-     *
-     * @return static A new instance of the current class with the specified link class.
-     */
-    public function linkClass(string $value): static
-    {
-        $new = clone $this;
-        CssClass::add($new->linkAttributes, $value);
 
         return $new;
     }
