+++ tests/Attribute/Component/HasLinkTest.php
@@ -9,37 +9,6 @@ use PHPUnit\Framework\TestCase;
 
 final class HasLinkTest extends TestCase
 {
-    public function testAttributes(): void
-    {
-        $instance = new class() {
-            use HasLink;
-
-            public function getLinkAttributes(): array
-            {
-                return $this->linkAttributes;
-            }
-        };
-
-        $instance = $instance->linkAttributes(['class' => 'foo']);
-        $instance = $instance->linkAttributes(['disabled' => true]);
-
-        $this->assertSame(['class' => 'foo', 'disabled' => true], $instance->getLinkAttributes());
-    }
-
-    public function testClass(): void
-    {
-        $instance = new class() {
-            use HasLink;
-
-            public function getLinkAttributes(): array
-            {
-                return $this->linkAttributes;
-            }
-        };
-
-        $this->assertSame(['class' => 'test-class'], $instance->linkClass('test-class')->getLinkAttributes());
-    }
-
     public function testGetLink(): void
     {
         $instance = new class() {
@@ -49,15 +18,6 @@ final class HasLinkTest extends TestCase
         $this->assertSame('test-link', $instance->link('test-link')->getLink());
     }
 
-    public function testGetLinkAttributes(): void
-    {
-        $instance = new class() {
-            use HasLink;
-        };
-
-        $this->assertSame(['class' => 'test-class'], $instance->linkAttributes(['class' => 'test-class'])->getLinkAttributes());
-    }
-
     public function testImmutablity(): void
     {
         $instance = new class() {
@@ -65,7 +25,5 @@ final class HasLinkTest extends TestCase
         };
 
         $this->assertNotSame($instance, $instance->link(''));
-        $this->assertNotSame($instance, $instance->linkAttributes([]));
-        $this->assertNotSame($instance, $instance->linkClass(''));
     }
 }
